\name{fastGeneMI-package}
\alias{fastGeneMI-package}
\alias{fastGeneMI}
\docType{package}
\title{
\packageTitle{fastGeneMI}
}
\description{
\packageDescription{fastGeneMI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastGeneMI}
\packageIndices{fastGeneMI}
Overview
}
\author{
\packageAuthor{fastGeneMI}

Maintainer: \packageMaintainer{fastGeneMI}
}
\references{
References
}
\keyword{ package }
\seealso{
Links
}

