% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get.mim.bspline}
\alias{get.mim.bspline}
\title{B-spline mutual information estimate}
\usage{
get.mim.bspline(expr.data, order, n.bins = as.integer(nrow(expr.data)^(1/3)),
  n.cores = 1L)
}
\arguments{
\item{expr.data}{A samples x genes matrix of expression values.}

\item{order}{The order of the B-spline used to smooth the histograms. This must be greater than 1 + the number of bins.}

\item{n.bins}{The number of bins into which the expression data will be discretised. The default is \eqn{N^{1/3}} for \eqn{N} samples.}

\item{n.cores}{The number of cores to use for the computation. The default is 1.}
}
\value{
A symmetric matrix of pairwise mutual information between the expression of pairs of genes. These are the mutual information estimtates between pairs of columns of \code{expr.data}.
}
\description{
Returns a symmetric matrix of pairwise B-spline mutual information estimates from a samples x genes matrix of expression values. The \eqn{ij^{th}} element is the mutual information estimate between the expression of genes \eqn{i} and \eqn{j}.
}
