% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get.mim.CS}
\alias{get.mim.CS}
\title{Chao-Shen Likelihood mutual information estimate}
\usage{
get.mim.CS(expr.data, discretisation = c("equalwidth", "equalfreq", "bb"),
  n.bins = as.integer(nrow(expr.data)^(1/3)), n.cores = 1L)
}
\arguments{
\item{expr.data}{A samples x genes matrix of expression values.}

\item{discretisation}{The types of bins into which the expression data will be discretised. Can choose equal width bins (\code{"equalwidth"}), equal frequency bins (\code{"equalfreq"}) or Bayesian Blocks bins (\code{"bb"})}

\item{n.bins}{The number of bins into which the expression data will be discretised. The default is \eqn{N^{1/3}} for \eqn{N} samples. Ignored if using Bayesian Blocks discretisation.}

\item{n.cores}{The number of cores to use for the computation. The default is 1.}
}
\value{
A symmetric matrix of pairwise mutual information between the expression of pairs of genes. These are the mutual information estimtates between pairs of columns of \code{expr.data}.
}
\description{
Returns a symmetric matrix of pairwise Chao-Shen mutual information estimates from a samples x genes matrix of expression values. The \eqn{ij^{th}} element is the mutual information estimate between the expression of genes \eqn{i} and \eqn{j}.
}
