% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get.mim.knn}
\alias{get.mim.knn}
\title{\eqn{k}-nearest-neighbour mutual information estimate}
\usage{
get.mim.knn(expr.data, k = 3, noise = 1e-09)
}
\arguments{
\item{expr.data}{A samples x genes matrix of expression values.}

\item{k}{The number of nearest neighbours to use. The default is 3}

\item{noise}{The amplitude of the random noise added to the expression values to break ties. Default value is 1e-09.}
}
\value{
A symmetric matrix of pairwise mutual information between the expression of pairs of genes. These are the mutual information estimtates between pairs of columns of \code{expr.data}.
}
\description{
Returns a symmetric matrix of pairwise \eqn{k}-nearest-neighbour mutual information estimates from a samples x genes matrix of expression values. The \eqn{ij^{th}} element is the mutual information estimate between the expression of genes \eqn{i} and \eqn{j}. This is a wrapper for the \code{parmigene} function \code{knnmi.all}.
}
