% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_inf.R
\name{get.pr}
\alias{get.pr}
\title{Compute the precision and recall from a matrix of mutual information estimates}
\usage{
get.pr(mim, gs.net, inf.algo = c("clr", "mrnet", "aracne"), n.reg = NULL)
}
\arguments{
\item{mim}{A symmetric matrix of mutual information values. The \eqn{ij^{th}} element is the mutual information between the expression of genes \eqn{i} and \eqn{j}}

\item{gs.net}{The gold standard regulatory network}

\item{inf.algo}{The inference algorithm used to infer the network from the mutual information estimates. Must be one of \code{"clr"}, \code{"mrnet"} or \code{"aracne"}}

\item{n.reg}{The number of genes that are designated as potential regulators. If it is an integer then the first \code{n.reg} genes are marked as regulators. If a vector then those gene indexes contained in the vector are regulators. Only interactions involving regulators are used to evaluate the precision and recall. Default isi \code{NULL}, in which case all the genes are potential regulators.}
}
\value{
ADD DOCUMENTATION HERE
}
\description{
From a symmetric matrix of mutual information values between the expression of pairs of genes, infer a network then compute the precision, recall and the area under precision-recall curve by comparing the inferred network to the gold standard
}
